requirements.txt
# Core data stack
numpy>=1.26
pandas>=2.2
pyarrow>=15.0

# Machine learning
scikit-learn>=1.4
joblib>=1.3

# Notebooks and exploration
jupyter>=1.0
ipykernel>=6.29
matplotlib>=3.8
seaborn>=0.13
plotly>=5.20

# Data acquisition
requests>=2.31
feedparser>=6.0
beautifulsoup4>=4.12
lxml>=5.1
python-dotenv>=1.0
tqdm>=4.66

# Dashboard
streamlit>=1.32
folium>=0.16
streamlit-folium>=0.20

# Testing and quality checks
pytest>=8.0
ruff>=0.3
